requirements.txt
python-rapidjson
requests
pyqt5
patool
black
